+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/ProcessEntry.java
@@ -27,6 +27,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.walmartlabs.concord.server.org.triggers.TriggerEntry;
 import com.walmartlabs.concord.server.sdk.ProcessStatus;
 import org.immutables.value.Value;
 
@@ -127,6 +128,9 @@ public interface ProcessEntry extends Serializable {
 
     @Nullable
     List<ProcessStatusHistoryEntry> statusHistory();
+
+    @Nullable
+    TriggeredByEntry triggeredBy();
 
     @Value.Immutable
     @JsonInclude(Include.NON_EMPTY)